refactor: Replace a popup with a form in sidebar
In the previous commits, to create a todo, you must first press a
dedicated button and type in your desired title in a popup form and then
submit it.

This commit removes the unnecessary(subjective) step which is having a
popup form. The user will have to just type in their desired title and
press 'enter' to create and save the todo.